requirements-omi.txt
# Omi Song Recognition Webhook Requirements

# Web framework
fastapi>=0.104.0
uvicorn[standard]>=0.24.0

# HTTP client for notifications
httpx>=0.25.0
requests>=2.31.0

# Environment variables
python-dotenv>=1.0.0

# Audio processing (already installed with abracadabra)
# numpy
# scipy
# pydub

# Abracadabra dependencies
# See main requirements.txt
